add server class hash map insert

+++ jay_irc/inc/server.hpp
@@ -35,7 +35,7 @@ private:
 	const std::string	__password;
 	int					__port_int;
 	std::map<int, Channel> *__channels;
-    std::map<unsigned long long, void(*)(Message &)> __cmd_list;
+    std::map<unsigned long, void(*)(Message &)> __cmd_list;
 	std::vector<Client> *__clients;
 };
